kg_backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
community==1.0.0b1
cssselect==1.2.0
distro==1.9.0
Django==5.1.2
django-cors-headers==4.6.0
django-ninja==1.3.0
docstring_parser==0.16
exceptiongroup==1.2.2
fastnode2vec==0.0.7
Flask==3.0.3
frozenlist==1.5.0
gensim==4.3.3
git-filter-repo==2.45.0
goose3==3.1.19
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
igraph==0.11.8
iniconfig==2.0.0
instructor==1.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
langdetect==1.0.9
leidenalg==0.10.2
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
node2vec==0.5.0
numba==0.60.0
numpy==1.26.4
openai==1.52.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
pyahocorasick==2.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-igraph==0.11.8
python-louvain==0.16
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.0.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.16.0
